feat(add event): rudimentary "add event" functionality
There are 3 types of events: finished, reading, dropped.
This commit allows users to create events by the following rules.
For each pair (user, book):
- there can be at most one event of each type;
- there cannot be both events of type dropped and finished;
- event of type reading must not have a later date than event of type
  finished/dropped.

Creating an event which violates any of the first 2 rules, replaces the
preexisting events.

Creating an event which violates the 3rd rule results in an internal
server error.